Moved extract_jobs_from_tron_yaml to utils to avoid circular imports
